--- common/src/main/kotlin/de/themeparkcraft/audioserver/common/interfaces/RabbitSendable.kt
@@ -1,16 +1,0 @@
-package de.themeparkcraft.audioserver.common.interfaces
-
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.encodeToByteArray
-import kotlinx.serialization.protobuf.ProtoBuf
-
-@Serializable
-sealed interface RabbitSendable {
-
-    @OptIn(ExperimentalSerializationApi::class)
-    fun encode(): ByteArray {
-        return ProtoBuf.encodeToByteArray(this)
-    }
-
-}
