--- common/src/main/kotlin/de/themeparkcraft/audioserver/common/rabbit/Constants.kt
@@ -1,4 +1,0 @@
-package de.themeparkcraft.audioserver.common.rabbit
-
-const val EXCHANGE_NAME = "themeparkcraft"
-const val ROUTIING_KEY = "themeparkcraft:audioserver"
