+++ common/src/main/kotlin/net/blockventuremc/audioserver/common/interfaces/AudioSource.kt
@@ -1,0 +1,14 @@
+package net.blockventuremc.audioserver.common.interfaces
+
+import kotlinx.serialization.Serializable
+
+/**
+ * The AudioSource interface represents an audio source. It provides access to the unique identifier (uid)
+ * and volume of the audio source.
+ */
+@Serializable
+sealed interface AudioSource : RabbitSendable {
+    val uid: String
+    val server: String
+    val volume: Float
+}
