+++ common/src/main/kotlin/net/blockventuremc/audioserver/common/interfaces/RabbitSendable.kt
@@ -1,0 +1,16 @@
+package net.blockventuremc.audioserver.common.interfaces
+
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.encodeToByteArray
+import kotlinx.serialization.protobuf.ProtoBuf
+
+@Serializable
+sealed interface RabbitSendable {
+
+    @OptIn(ExperimentalSerializationApi::class)
+    fun encode(): ByteArray {
+        return ProtoBuf.encodeToByteArray(this)
+    }
+
+}
