+++ common/src/main/kotlin/net/blockventuremc/audioserver/common/rabbit/Constants.kt
@@ -1,0 +1,4 @@
+package net.blockventuremc.audioserver.common.rabbit
+
+const val EXCHANGE_NAME = "venturelibs"
+const val ROUTIING_KEY = "venturelibs:audioserver"
