+++ gradle.properties
@@ -1,2 +1,2 @@
 kotlin.code.style=official
-audioServerVersion=0.5-Snapshot
+audioServerVersion=2024.05.07-SNAPSHOT
