--- standalone/src/main/kotlin/de/themeparkcraft/audioserver/AudioServerStandalone.kt
@@ -1,46 +1,0 @@
-package de.themeparkcraft.audioserver
-
-import de.themeparkcraft.audioserver.common.data.RabbitConfiguration
-import de.themeparkcraft.audioserver.common.extensions.getLogger
-import de.themeparkcraft.audioserver.common.interfaces.PlayerPositionUpdate
-import de.themeparkcraft.audioserver.common.rabbit.RabbitClient
-import de.themeparkcraft.audioserver.utils.Environment
-import kotlin.system.exitProcess
-
-class AudioServerStandalone {
-
-    private val rabbitClient: RabbitClient
-
-    init {
-        val rabbitConfiguration = RabbitConfiguration(
-            host = Environment.getEnv("RABBITMQ_HOST") ?: "localhost",
-            port = Environment.getEnv("RABBITMQ_PORT")?.toInt() ?: 5672,
-            virtualHost = Environment.getEnv("RABBITMQ_VIRTUAL_HOST") ?: "/",
-            username = Environment.getEnv("RABBITMQ_USER") ?: "guest",
-            password = Environment.getEnv("RABBITMQ_PASSWORD") ?: "guest"
-        )
-
-        getLogger().info("Starting AudioServer standalone at ${rabbitConfiguration.host}:${rabbitConfiguration.port} - (vhost: ${rabbitConfiguration.virtualHost}) with user ${rabbitConfiguration.username}...")
-
-        try {
-            rabbitClient = RabbitClient(rabbitConfiguration)
-
-            rabbitClient.withListener { message, delivery ->
-                when (val rabbitSendable = rabbitClient.deserializeRabbitMessage(delivery.body)) {
-                    is PlayerPositionUpdate -> {
-                        getLogger().info("Received PlayerPositionUpdate message: ${rabbitSendable.playerUid} at ${rabbitSendable.position}")
-                    }
-                    else -> {
-                        getLogger().warn("Received unknown RabbitSendable message: $rabbitSendable")
-                    }
-                }
-            }
-
-            getLogger().info("AudioServer standalone started.")
-        } catch (exception: Exception) {
-            getLogger().error("Failed to start AudioServer standalone due to the following exception: ", exception)
-            exitProcess(1)
-        }
-    }
-
-}
