--- standalone/src/main/kotlin/de/themeparkcraft/audioserver/Start.kt
@@ -1,5 +1,0 @@
-package de.themeparkcraft.audioserver
-
-fun main() {
-    AudioServerStandalone()
-}
