+++ standalone/src/main/kotlin/net/blockventuremc/audioserver/Start.kt
@@ -1,0 +1,5 @@
+package net.blockventuremc.audioserver
+
+fun main() {
+    AudioServerStandalone()
+}
